requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
Jinja2==3.1.3
MarkupSafe==2.1.5
click==8.1.7
itsdangerous==2.1.2
python-dotenv==1.0.0
PyJWT==2.8.0
python-jose==3.3.0
supabase==2.3.4
postgrest==0.13.1
httpx==0.26.0 